scripts/web-app/backend/requirements.txt
absl-py==1.4.0
addict==2.4.0
aiofiles==0.7.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
altair==4.2.2
anyio==3.6.1
apache-beam==2.46.0
argcomplete==1.11.0
arrow==1.2.3
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
avro-python3==1.10.2
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.cached-property==1.0.2
beautifulsoup4==4.12.0
black==23.1.0
blessed==1.20.0
boto3==1.17.36
botocore==1.20.112
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.0.4
cloudpickle==2.2.1
coco-assistant==0.3.5
COCO-merger==0.0.1
colorama==0.4.5
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.0.5
crcmod==1.7
croniter==1.3.8
cycler==0.11.0
Cython==0.29.32
dacite==1.6.0
dateutils==0.6.12
debugpy @ file:///tmp/build/80754af9/debugpy_1637091799509/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
deepdiff==6.3.0
Deprecated==1.2.11
dill==0.3.1.1
dm-tree==0.1.8
dnspython==1.16.0
docopt==0.6.2
efficientnet==1.0.0
efficientnet-pytorch==0.7.1
einops==0.6.0
email-validator==1.3.1
ensemble-boxes==1.0.9
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926439650/work
et-xmlfile==1.1.0
etils==1.1.1
eventlet==0.31.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.88.0
fastavro==1.7.3
fasteners==0.18
ffmpy==0.3.0
fiftyone==0.17.2
fiftyone-brain==0.9.1
fiftyone-db==0.3.0
fiftyone-desktop==0.23.2
filelock==3.12.0
fire==0.4.0
Flask==1.1.2
flatbuffers==1.12
fonttools==4.37.3
frozenlist==1.3.3
fsspec==2023.1.0
funcy==1.17
future==0.18.2
gast==0.4.0
gin-config==0.5.0
glob2==0.7
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
gradio==3.19.1
graphql-core==3.1.7
greenlet==1.1.3
grpcio==1.51.3
h11==0.12.0
h2==4.1.0
h5py==3.8.0
hdfs==2.7.0
hpack==4.0.0
httpcore==0.15.0
httplib2==0.21.0
httptools==0.5.0
httpx==0.23.0
huggingface-hub==0.13.4
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.4
image-classifiers==1.0.0
imagecodecs==2023.3.16
imageio==2.22.1
ImgAnn==0.8.1
imgaug==0.4.0
immutabledict==2.2.3
importlib-metadata==4.12.0
importlib-resources==5.12.0
inquirer==3.1.3
ipykernel @ file:///opt/conda/conda-bld/ipykernel_1662361798230/work
ipython @ file:///opt/conda/conda-bld/ipython_1657652213665/work
itsdangerous==2.1.2
jax==0.4.8
jedi @ file:///tmp/build/80754af9/jedi_1644297102865/work
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.2.0
jsonschema==4.17.0
jupyter-core @ file:///opt/conda/conda-bld/jupyter_core_1651671229925/work
jupyter_client @ file:///opt/conda/conda-bld/jupyter_client_1662504365333/work
kaggle==1.5.13
kaleido==0.2.1
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
labelme2coco==0.2.1
libclang==15.0.6.1
lightning==2.0.0
lightning-cloud==0.5.32
lightning-utilities==0.8.0
linkify-it-py==2.0.0
llvmlite==0.39.1
loguru==0.5.3
lvis==0.5.3
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.3.4
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdit-py-plugins==0.3.3
mdurl==0.1.2
ml-dtypes==0.1.0
mmcv-full==1.6.1
# Editable install with no version control (mmdet==2.25.3)
-e /home/badboy-002/anaconda3/envs/senior_env/lib/python3.9/site-packages
model-index==0.1.11
mongoengine==0.20.0
motor==2.3.0
multidict==6.0.4
munch==2.5.0
mypy-extensions==0.4.3
natsort==8.3.1
ndjson==0.3.1
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1649847906199/work
networkx==2.8.7
numba==0.56.4
numcodecs==0.11.0
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cudnn-cu11==8.6.0.163
oauth2client==4.1.3
oauthlib==3.2.2
object-detection @ file:///home/badboy-002/github/senior_project/web-app/backend/models/research
objsize==0.6.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openmim==0.3.2
openpyxl==3.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.8.6
packaging==23.0
pandas==1.2.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathspec==0.11.0
patool==1.12
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==2.6.2
plotly==5.2.2
portalocker==2.7.0
pprintpp==0.4.0
pretrainedmodels==0.7.4
priority==2.0.0
promise==2.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
proto-plus==1.22.2
protobuf==3.19.6
psutil==5.7.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-cpuinfo==9.0.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.5
pycocotools==2.0.4
pycryptodome==3.17
pydantic==1.10.5
pydot==1.4.2
pydub==0.25.1
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PyJWT==2.6.0
pymongo==3.11.0
pyodi==0.0.7
pyparsing==2.4.7
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.19.2
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.5
python-slugify==8.0.1
python-version==0.0.2
pytorch-lightning==2.0.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==5.4
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
qudida==0.0.4
readchar==4.0.5
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
retrying==1.3.3
rfc3986==1.5.0
rich==12.5.1
rsa==4.9
s3transfer==0.3.7
sacrebleu==2.2.0
sahi==0.11.1
scikit-image==0.19.3
scikit-learn==1.2.2
scikit-multilearn==0.2.0
scipy==1.9.1
seaborn==0.12.1
segmentation-models==1.0.1
segmentation-models-pytorch @ git+https://github.com/qubvel/segmentation_models.pytorch@e7cbea0bbd546167b9d1757b2cfeca3fbe26a33e
Send2Trash==1.8.0
sentencepiece==0.1.97
sentinel==0.3.0
seqeval==1.2.2
Shapely==1.8.5.post1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
split-folders==0.5.1
sse-starlette==0.10.3
sseclient-py==1.7.2
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.22.0
starsessions==1.3.0
strawberry-graphql==0.96.0
tabulate==0.8.10
tenacity==8.1.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.9.1
tensorflow-addons==0.19.0
tensorflow-datasets==4.8.3
tensorflow-estimator==2.9.0
tensorflow-hub==0.13.0
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.12.0
tensorflow-model-optimization==0.7.3
tensorflow-text==2.11.0
termcolor==2.1.0
terminaltables==3.1.10
text-unidecode==1.3
tf-models-official==2.11.4
tf-slim==1.1.0
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.6.13
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchmetrics==0.11.0
torchsummary==1.5.1
torchvision==0.13.1+cu116
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.64.1
traitlets==5.9.0
typeguard==3.0.1
typing_extensions==4.5.0
tzdata==2022.4
tzlocal==4.2
uc-micro-py==1.0.1
ujson==5.7.0
universal-analytics-python3==1.0.1
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.20.0
uvloop==0.17.0
voxel51-eta==0.8.1
watchfiles==0.19.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
xmltodict==0.12.0
y-py==0.5.9
yapf==0.32.0
yarg==0.1.9
yarl==1.8.2
zipp==3.8.1
zstandard==0.20.0
